:bug: Empty filter list no longer breaks list
Userlist was formerly checking if the currently filtered list had users in it to display the whole list component, including the fitler buttons. Caused issue where buttons and list would disappear entirely when a filter was applied that returned no users. Now the list displays so long as any users are blocked OR allowed, and if the current filtered list is empty it displays a message to that effect and prompts the user to check out the other filter options